requirements.txt
torch
torchvision
ultralytics
opencv-python
pyserial
numpy
tensorrt
pycuda
